+++ halfapi/conf.py
@@ -46,19 +46,51 @@ import uuid
 
 import toml
 
-PRODUCTION = True
-LOGLEVEL = 'info'
-CONF_FILE = os.environ.get('HALFAPI_CONF_FILE', '.halfapi/config')
-DRYRUN = bool(os.environ.get('HALFAPI_DRYRUN', False))
-
 SCHEMA = {}
 
-CONF_DIR = environ.get('HALFAPI_CONF_DIR', '/etc/half_api')
+DEFAULT_CONF = {
+    # Default configuration values
+    'SECRET': tempfile.mkstemp()[1],
+    'PROJECT_NAME': os.getcwd().split('/')[-1],
+    'PRODUCTION': True,
+    'HOST': '127.0.0.1',
+    'PORT': 3000,
+    'LOGLEVEL': 'info',
+    'BASE_DIR': os.getcwd(),
+    'CONF_FILE': '.halfapi/config',
+    'CONF_DIR': '/etc/half_api',
+    'DRYRUN': None
+}
+
+PROJECT_LEVEL_KEYS = {
+    # Allowed keys in "project" section of configuration file
+    'project_name',
+    'production',
+    'secret',
+    'host',
+    'port',
+    'loglevel',
+    'dryrun'
+}
+
+DOMAIN_LEVEL_KEYS = PROJECT_LEVEL_KEYS | {
+    # Allowed keys in "domain" section of configuration file
+    'name',
+    'module',
+    'prefix',
+    'enabled'
+}
+
+CONF_FILE = os.environ.get('HALFAPI_CONF_FILE', DEFAULT_CONF['CONF_FILE'])
+CONF_DIR = os.environ.get('HALFAPI_CONF_DIR', DEFAULT_CONF['CONF_DIR'])
+
 HALFAPI_ETC_FILE=os.path.join(
     CONF_DIR, 'config'
 )
+
+BASE_DIR = os.environ.get('HALFAPI_BASE_DIR', DEFAULT_CONF['BASE_DIR'])
 HALFAPI_DOT_FILE=os.path.join(
-    os.getcwd(), '.halfapi', 'config')
+    BASE_DIR, '.halfapi', 'config')
 
 HALFAPI_CONFIG_FILES = []
 
@@ -66,14 +98,35 @@ try:
     with open(HALFAPI_ETC_FILE, 'r'):
         HALFAPI_CONFIG_FILES.append(HALFAPI_ETC_FILE)
 except FileNotFoundError:
-    logger.error('Cannot find a configuration file under %s', HALFAPI_DOT_FILE)
+    logger.info('Cannot find a configuration file under %s', HALFAPI_ETC_FILE)
 
 try:
     with open(HALFAPI_DOT_FILE, 'r'):
         HALFAPI_CONFIG_FILES.append(HALFAPI_DOT_FILE)
 except FileNotFoundError:
-    logger.error('Cannot find a configuration file under %s', HALFAPI_DOT_FILE)
+    logger.info('Cannot find a configuration file under %s', HALFAPI_DOT_FILE)
 
+
+ENVIRONMENT = {}
+# Load environment variables allowed in configuration
+
+if 'HALFAPI_DRYRUN' in os.environ:
+    ENVIRONMENT['dryrun'] = True
+
+if 'HALFAPI_PROD' in os.environ:
+    ENVIRONMENT['production'] = bool(os.environ.get('HALFAPI_PROD'))
+
+if 'HALFAPI_LOGLEVEL' in os.environ:
+    ENVIRONMENT['loglevel'] = os.environ.get('HALFAPI_LOGLEVEL').lower()
+
+if 'HALFAPI_SECRET' in os.environ:
+    ENVIRONMENT['secret'] = os.environ.get('HALFAPI_SECRET')
+
+if 'HALFAPI_HOST' in os.environ:
+    ENVIRONMENT['host'] = os.environ.get('HALFAPI_HOST')
+
+if 'HALFAPI_PORT' in os.environ:
+    ENVIRONMENT['port'] = int(os.environ.get('HALFAPI_PORT'))
 
 def read_config(filenames=HALFAPI_CONFIG_FILES):
     """
@@ -84,18 +137,22 @@ def read_config(filenames=HALFAPI_CONFIG_FILES):
     logger.info('Reading config files %s', filenames)
     for CONF_FILE in filenames:
         if os.path.isfile(CONF_FILE):
-            d_res.update( toml.load(CONF_FILE) )
+            conf_dict = toml.load(CONF_FILE)
+            d_res.update(conf_dict)
 
     logger.info('Read config files (result) %s', d_res)
     return { **d_res.get('project', {}), 'domain': d_res.get('domain', {}) }
 
 CONFIG = read_config()
+CONFIG.update(**ENVIRONMENT)
 
 PROJECT_NAME = CONFIG.get('project_name',
-    environ.get('HALFAPI_PROJECT_NAME', os.getcwd().split('/')[-1]))
+    os.environ.get('HALFAPI_PROJECT_NAME', DEFAULT_CONF['PROJECT_NAME']))
 
-if environ.get('HALFAPI_DOMAIN_NAME'):
-    DOMAIN_NAME = environ.get('HALFAPI_DOMAIN_NAME')
+if os.environ.get('HALFAPI_DOMAIN_NAME'):
+    # Force enabled domain by environment variable
+
+    DOMAIN_NAME = os.environ.get('HALFAPI_DOMAIN_NAME')
     if 'domain' in CONFIG and DOMAIN_NAME in CONFIG['domain'] \
         and 'config' in CONFIG['domain'][DOMAIN_NAME]:
 
@@ -113,53 +170,36 @@ if environ.get('HALFAPI_DOMAIN_NAME'):
 
     CONFIG['domain'][DOMAIN_NAME]['config'] = domain_config
 
-    if environ.get('HALFAPI_DOMAIN_MODULE'):
-        dom_module = environ.get('HALFAPI_DOMAIN_MODULE')
+    if os.environ.get('HALFAPI_DOMAIN_MODULE'):
+        # Specify the pythonpath to import the specified domain (defaults to global)
+        dom_module = os.environ.get('HALFAPI_DOMAIN_MODULE')
         CONFIG['domain'][DOMAIN_NAME]['module'] = dom_module
 
 if len(CONFIG.get('domain', {}).keys()) == 0:
     logger.info('No domains')
 
-# Bind
-HOST = CONFIG.get('host',
-    environ.get('HALFAPI_HOST', '127.0.0.1'))
-PORT = int(CONFIG.get(
-    'port',
-    environ.get('HALFAPI_PORT', '3000')))
-
 
 # Secret
-SECRET = CONFIG.get(
-    'secret',
-    environ.get('HALFAPI_SECRET'))
-
-if not SECRET:
+if 'secret' not in CONFIG:
     # TODO: Create a temporary secret
-    _, SECRET = tempfile.mkstemp()
-    with open(SECRET, 'w') as secret_file:
+    CONFIG['secret'] = DEFAULT_CONF['SECRET']
+    with open(CONFIG['secret'], 'w') as secret_file:
         secret_file.write(str(uuid.uuid4()))
 
 try:
-    with open(SECRET, 'r') as secret_file:
-        CONFIG['secret'] = SECRET.strip()
+    with open(CONFIG['secret'], 'r') as secret_file:
+        CONFIG['secret'] = CONFIG['secret'].strip()
 except FileNotFoundError as exc:
-    logger.info('Running without secret file: %s', SECRET or 'no file specified')
+    logger.warning('Running without secret file: %s', CONFIG['secret'] or 'no file specified')
 
-PRODUCTION = bool(CONFIG.get(
-    'production',
-    environ.get('HALFAPI_PROD', True)))
+CONFIG.setdefault('project_name', DEFAULT_CONF['PROJECT_NAME'])
+CONFIG.setdefault('production', DEFAULT_CONF['PRODUCTION'])
+CONFIG.setdefault('host', DEFAULT_CONF['HOST'])
+CONFIG.setdefault('port', DEFAULT_CONF['PORT'])
+CONFIG.setdefault('loglevel', DEFAULT_CONF['LOGLEVEL'])
+CONFIG.setdefault('dryrun', DEFAULT_CONF['DRYRUN'])
 
-LOGLEVEL = CONFIG.get(
-    'loglevel',
-    environ.get('HALFAPI_LOGLEVEL', 'info')).lower()
-
-BASE_DIR = CONFIG.get(
-    'base_dir',
-    environ.get('HALFAPI_BASE_DIR', '.'))
-
-CONFIG['project_name'] = PROJECT_NAME
-CONFIG['production'] = PRODUCTION
-CONFIG['secret'] = SECRET
-CONFIG['host'] = HOST
-CONFIG['port'] = PORT
-CONFIG['dryrun'] = DRYRUN
+# !!!TO REMOVE!!!
+SECRET = CONFIG.get('SECRET')
+PRODUCTION = CONFIG.get('production')
+# !!!