+++ halfapi/logging.py
@@ -1,8 +1,10 @@
 import logging
 
+default_level = logging.DEBUG
+default_format = '%(asctime)s [%(process)d] [%(levelname)s] %(message)s'
+default_datefmt = '[%Y-%m-%d %H:%M:%S %z]'
 
-def config_logging(level=logging.INFO):
-
+def config_logging(level=default_level, format=default_format, datefmt=default_datefmt):
     # When run by 'uvicorn ...', a root handler is already
     # configured and the basicConfig below does nothing.
     # To get the desired formatting:
@@ -12,8 +14,8 @@ def config_logging(level=logging.INFO):
     #   https://github.com/encode/uvicorn/issues/511
     logging.basicConfig(
         # match gunicorn format
-        format='%(asctime)s [%(process)d] [%(levelname)s] %(message)s',
-        datefmt='[%Y-%m-%d %H:%M:%S %z]',
+        format=format,
+        datefmt=datefmt,
         level=level)
 
         # When run by 'gunicorn -k uvicorn.workers.UvicornWorker ...',
@@ -27,5 +29,4 @@ def config_logging(level=logging.INFO):
     logging.getLogger('uvicorn.access').propagate = True
     logging.getLogger('uvicorn.error').propagate = True
 
-config_logging()
 logger = logging.getLogger()